Moved file shuffler functionality to shuffler_api.py
